pyproject.toml
[tool.poetry]
name = "grpc-asgiref"
version = "0.1.0"
description = ""
authors = ["Charlie DeTar <charlie@literati.com>"]

[tool.poetry.dependencies]
python = "^3.9"
grpcio = "^1.44.0"
asgiref = "^3.5.0"

[tool.poetry.dev-dependencies]
grpcio-tools = "^1.44.0"
pytest = "^7.1.1"
pytest-xdist = "^2.5.0"
pytest-asyncio = "^0.18.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
